requirements/requirements_webapp.txt
streamlit==1.25.0
grpcio==1.56.2